pipeline/kafka_pipeline/requirements.txt
python-dotenv
pandas
confluent_kafka
psycopg2-binary